% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{arctic}
\alias{arctic}
\title{SpatialPolygonsDataFrame containing land shapes for pan-Arctic map with a cutpoint at 40 degrees latitude.}
\format{
SpatialPolygonsDataFrame
}
\source{
\href{http://www.naturalearthdata.com/}{Natural Earth Data}
}
\usage{
arctic
}
\description{
SpatialPolygonsDataFrame containing land shapes for pan-Arctic map with a cutpoint at 40 degrees latitude.
}
\seealso{
basemap

Other shapefiles: 
\code{\link{arctic60}},
\code{\link{arctic_bathy}},
\code{\link{barents_bathy}},
\code{\link{barents_currents}},
\code{\link{kong.cr}},
\code{\link{mosj.cr}},
\code{\link{svalbard.gl}},
\code{\link{svalbard.ld}},
\code{\link{svalbard_bathy}}
}
\concept{shapefiles}
\keyword{datasets}
\keyword{internal}
\keyword{shapefiles}
