% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{arctic50}
\alias{arctic50}
\title{SpatialPolygonsDataFrame containing land shapes for pan-Arctic map with a cutpoint at 50 degrees latitude.}
\format{SpatialPolygonsDataFrame}
\source{
\href{http://www.naturalearthdata.com/}{Natural Earth Data}
}
\usage{
arctic50
}
\description{
SpatialPolygonsDataFrame containing land shapes for pan-Arctic map with a cutpoint at 50 degrees latitude.
}
\seealso{
basemap

Other shapefiles: \code{\link{arctic60}},
  \code{\link{arctic_bathy}},
  \code{\link{arctic_currents}},
  \code{\link{atlantic_currents}},
  \code{\link{barents_bathy}}, \code{\link{kong.cr}},
  \code{\link{mosj.cr}}, \code{\link{svalbard.gl}},
  \code{\link{svalbard.ld}}, \code{\link{svalbard_bathy}}
}
\concept{shapefiles}
\keyword{datasets}
\keyword{internal}
\keyword{shapefiles}
