% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{arctic60}
\alias{arctic60}
\alias{barents.ld}
\title{SpatialPolygonsDataFrame containing land shapes for the Barents Sea region.}
\format{SpatialPolygonsDataFrame}
\source{
\href{http://www.naturalearthdata.com/}{Natural Earth Data}
}
\usage{
barents.ld
}
\description{
SpatialPolygonsDataFrame containing land shapes for the Barents Sea region.
}
\seealso{
basemap

Other shapefiles: \code{\link{arctic_bathy}},
  \code{\link{arctic}}, \code{\link{barents_bathy}},
  \code{\link{barents_currents}}, \code{\link{kong.cr}},
  \code{\link{mosj.cr}}, \code{\link{svalbard.gl}},
  \code{\link{svalbard.ld}}, \code{\link{svalbard_bathy}}
}
\concept{shapefiles}
\keyword{datasets}
\keyword{internal}
\keyword{shapefiles}
