% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{arctic_bathy}
\alias{arctic_bathy}
\title{SpatialPolygonsDataFrame containing pan-Arctic bathymetry}
\format{SpatialPolygonsDataFrame containing pan-Arctic bathymetry. In polar stereographic 71 degrees projection. Generalized from GEBCO One Minute Grid}
\source{
\href{https://www.gebco.net/data_and_products/gridded_bathymetry_data/gebco_one_minute_grid/}{General Bathymetric Chart of the Oceans}
}
\usage{
arctic_bathy
}
\description{
SpatialPolygonsDataFrame containing pan-Arctic bathymetry
}
\seealso{
basemap

Other shapefiles: \code{\link{arctic60}},
  \code{\link{arctic}}, \code{\link{barents_bathy}},
  \code{\link{barents_currents}}, \code{\link{kong.cr}},
  \code{\link{mosj.cr}}, \code{\link{svalbard.gl}},
  \code{\link{svalbard.ld}}, \code{\link{svalbard_bathy}}
}
\concept{shapefiles}
\keyword{datasets}
\keyword{internal}
\keyword{shapefiles}
