% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{atlantic_currents}
\alias{atlantic_currents}
\title{Data frame containing Atlantic ocean currents in the Barents Sea}
\format{Data frame}
\source{
Institute of Marine Research, Mikko Vihtakari
}
\usage{
atlantic_currents
}
\description{
Data frame containing Atlantic ocean currents in the Barents Sea
}
\seealso{
basemap

Other shapefiles: \code{\link{arctic50}},
  \code{\link{arctic60}}, \code{\link{arctic_bathy}},
  \code{\link{arctic_currents}},
  \code{\link{barents_bathy}}, \code{\link{kong.cr}},
  \code{\link{mosj.cr}}, \code{\link{svalbard.ld}},
  \code{\link{svalbard_bathy}}
}
\concept{shapefiles}
\keyword{datasets}
\keyword{internal}
\keyword{shapefiles}
