% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{barents_bathy}
\alias{barents_bathy}
\title{SpatialPolygonsDataFrame containing bathymetry for the Barents Sea}
\format{SpatialPolygonsDataFrame containing pan-Arctic bathymetry. In \code{"+init=epsg:32633"} projection.}
\source{
\href{https://www.ngdc.noaa.gov/mgg/bathymetry/arctic/ibcaoversion3.html}{Jakobsson, M., et al. The International Bathymetric Chart of the Arctic Ocean (IBCAO) Version 3.0. Geophys. Res. Lett. 2012, 39:L12609}
}
\usage{
barents_bathy
}
\description{
SpatialPolygonsDataFrame containing bathymetry for the Barents Sea
}
\seealso{
basemap

Other shapefiles: \code{\link{arctic60}},
  \code{\link{arctic_bathy}}, \code{\link{arctic}},
  \code{\link{barents_currents}}, \code{\link{kong.cr}},
  \code{\link{mosj.cr}}, \code{\link{svalbard.gl}},
  \code{\link{svalbard.ld}}, \code{\link{svalbard_bathy}}
}
\concept{shapefiles}
\keyword{datasets}
\keyword{internal}
\keyword{shapefiles}
