% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{barents_currents}
\alias{barents_currents}
\title{Data frame containing Atlantic and Arctic ocean currents for the Barents Sea}
\format{Data frame}
\source{
Institute of Marine Research (mostly Harald Gjøsæter), Norwegian Polar Institute (Arild Sundfjord, Mikko Vihtakari).

Modified from Eriksen E, Gjøsæter H, Prozorkevich D et al. (2018) From single species surveys towards monitoring of the Barents Sea ecosystem. Progress in Oceanography, 166, 4–14. doi:https://doi.org/10.1016/j.pocean.2017.09.007
}
\usage{
barents_currents
}
\description{
The data frame, intended to be shown as arrows, represents the current understanding of Atlantic and Arctic ocean currents. The currents are updated as our knowledge increases. The Barent Sea currents are based on Eriksen et al. 2018, while North of Svalbard currents are based on the articles published by Arild Sundfjord.
}
\seealso{
basemap

Other shapefiles: \code{\link{arctic50}},
  \code{\link{arctic60}}, \code{\link{arctic_bathy}},
  \code{\link{barents_bathy}}, \code{\link{kong.cr}},
  \code{\link{mosj.cr}}, \code{\link{svalbard.gl}},
  \code{\link{svalbard.ld}}, \code{\link{svalbard_bathy}}
}
\concept{shapefiles}
\keyword{datasets}
\keyword{internal}
\keyword{shapefiles}
