% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap_data.R
\name{basemap_data}
\alias{basemap_data}
\title{Create basemapData object for basemap plotting}
\usage{
basemap_data(
  type,
  limits = NULL,
  round.lat. = round.lat,
  n.lat.grid. = n.lat.grid,
  round.lon. = round.lon,
  n.lon.grid. = n.lon.grid,
  lat.interval. = lat.interval,
  lon.interval. = lon.interval,
  expar = 0,
  clip = TRUE,
  keep.glaciers. = keep.glaciers
)
}
\arguments{
\item{type}{See \code{\link{basemap}}}

\item{limits}{See \code{\link{basemap}}}

\item{round.lat.}{See \code{\link{basemap}}}

\item{n.lat.grid.}{See \code{\link{basemap}}}

\item{round.lon.}{See \code{\link{basemap}}}

\item{n.lon.grid.}{See \code{\link{basemap}}}

\item{lat.interval.}{See \code{\link{basemap}}}

\item{lon.interval.}{See \code{\link{basemap}}}

\item{expar}{Expansion parameter for \code{\link{deg_grid}}}

\item{clip}{Logical. Should shapefiles be clipped before plotting?}

\item{keep.glaciers.}{See \code{\link{basemap}}}
}
\description{
Internal function to create a \code{basemapData} object for \code{\link{basemap}}
}
\details{
This is an internal function, which is automatically run by the \code{\link{basemap}} function. Common users do not need to worry about these details.
}
\seealso{
\code{\link{basemap}} \code{\link{deg_grid}} \code{\link{deg_grid_polar}}
}
\author{
Mikko Vihtakari
}
\keyword{internal}
