% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap_limits.R
\name{basemap_limits}
\alias{basemap_limits}
\title{Define limits for a basemap}
\usage{
basemap_limits(
  limits,
  proj_deg = "+proj=longlat +datum=WGS84",
  type = MapType$map.type
)
}
\arguments{
\item{limits}{See \code{\link{basemap}}}

\item{proj_deg}{Projection for \code{limits}. Defaults to decimal degrees}

\item{type}{Map type for output. Defines the used UTM projection (see \code{\link{basemap}} and \code{\link{map_type}})}
}
\description{
An internal function to define limits for a basemap
}
\details{
This is an internal function, which is automatically run by the \code{\link{basemap}} function. Common users do not need to worry about these details.
}
\seealso{
\code{\link{basemap}} \code{\link{basemap_data}} \code{\link{deg_grid}} \code{\link{deg_grid_polar}}
}
\author{
Mikko Vihtakari
}
\keyword{internal}
