% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{chlorophyll}
\alias{chlorophyll}
\title{Chlorophyll-a recordings in Kongsfjorden}
\format{
A dataframe containing following columns:
\itemize{
  \item Name. Sample name.
  \item Station. Sampling station.
  \item lon. Longitude of the sampling location in decimal degrees
  \item lat. Latitude of the sampling location in decimal degrees.
  \item lon.utm. Longitude of the sampling location as UTM coordinates (\code{"+init=epsg:32633"})
  \item lat.utm. Latitude of the sampling location as UTM coordinates (\code{"+init=epsg:32633"})
  \item From. Sampling depth (m)
  \item Chla. Chlorophyll-a value (mg/m3)
 }
}
\source{
Norwegian Polar Institute (\url{https://data.npolar.no/home/})
}
\usage{
data(chlorophyll)
}
\description{
Chlorophyll-a recordings in Kongsfjorden
}
\keyword{datasets}
