% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_bathymetry.R
\name{clip_bathymetry}
\alias{clip_bathymetry}
\title{Clip a bathymetry shapefile to fit a basemap}
\usage{
clip_bathymetry(X, detailed = FALSE)
}
\arguments{
\item{X}{A basemapData object}

\item{detailed}{Logical indicating whether detailed bathymetry shapefiles should be used for Svalbard maps.}
}
\description{
An internal utility function to clip bathymetry to fit the boundaries of a basemap
}
\author{
Mikko Vihtakari
}
\keyword{internal}
