% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_current.R
\name{clip_current}
\alias{clip_current}
\title{Clip ocean current data frame to fit a basemap}
\usage{
clip_current(dt, X, shrink = 0.5)
}
\arguments{
\item{dt}{data frame containing ocean currents. Must be run through the \code{\link[broom]{tidy}} function to contain \code{long}, \code{lat} and \code{group} columns.}

\item{X}{A basemapData object}

\item{shrink}{numeric between 0 and 100 defining the percentage of clipping ocean current arrows to be done vertically and horizontally.}
}
\description{
An internal utility function to clip ocean current data frames to fit the boundaries of a basemap
}
\author{
Mikko Vihtakari
}
\keyword{internal}
