% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{ctd_rijpfjord}
\alias{ctd_rijpfjord}
\title{Example CTD data for Rijpfjorden}
\format{data.frame}
\usage{
ctd_rijpfjord
}
\description{
A dataframe containing example CTD data for Rijpfjorden.
}
\seealso{
Other ts_plot: \code{\link{ctd_kongsfjord}},
  \code{\link{define_water_type}},
  \code{\link{kongsfjord_watermasses}},
  \code{\link{rijpfjord_watermasses}},
  \code{\link{ts_plot}}
}
\concept{ts_plot}
\keyword{datasets}
