% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_grid.R
\name{deg_grid}
\alias{deg_grid}
\title{Decimal degree grid from UTM projected spatial objects}
\usage{
deg_grid(
  dat,
  round.lat = FALSE,
  n.lat.grid = 3,
  round.lon = FALSE,
  n.lon.grid = 3,
  n.points = 30
)
}
\arguments{
\item{dat}{\code{\link[=basemap_limits]{basemapLimits}} object}

\item{round.lat}{specifying the level of rounding to be used to plot latitude grid lines. Overrides \code{n.lat.grid}}

\item{n.lat.grid}{number of latitude grid lines. Alternatively use \code{round.lat}}

\item{round.lon}{specifying the level of rounding to be used to plot longitude grid lines. Overrides \code{n.lon.grid}}

\item{n.lon.grid}{number of longitude grid lines. Alternatively use \code{round.lon}}

\item{n.points}{number of points used in creating the grid lines. The more points, the smoother the lines.}
}
\value{
Returns a UTM coodrinate grid.
}
\description{
Outputs a list of grid lines that can be used to plot a grid on maps. Requires a spatial object projected using UTM coordinates
}
\details{
This is an internal function, which is automatically run by the \code{\link{basemap}} function. Common users do not need to worry about these details.
}
\seealso{
\code{\link{basemap}}
}
\author{
Mikko Vihtakari
}
\keyword{internal}
