% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_grid_polar.R
\name{deg_grid_polar}
\alias{deg_grid_polar}
\title{Decimal degree grid from polar stereographic spatial objects}
\usage{
deg_grid_polar(dat, lat.interval = 10, lon.interval = 45,
  n.points = 1000, proj4.utm = TRUE,
  proj4.deg = "+proj=longlat +datum=WGS84")
}
\arguments{
\item{dat}{Spatial object projected using UTM coordinates, such as \code{\link{readShapeSpatial}} OR a numeric vector of length 4 where first element defines the minimum longitude, second element the maximum longitude, third element the minimum latitude and fourth element the maximum latitude of the bounding box.}

\item{lat.interval}{the interval of latitude grids for polar stereographic maps (\code{type = "arctic50"} or \code{"arctic60"})}

\item{lon.interval}{the interval of longitude grids for polar stereographic maps (\code{type = "arctic50"} or \code{"arctic60"})}

\item{n.points}{number of points used in creating the grid lines. The more points, the smoother the lines.}

\item{proj4.utm}{projection in \code{dat}. If TRUE (default), fetches the projection from \code{dat}. If \code{dat} is a numeric vector and \code{proj4.utm = TRUE}, \code{CRS("+init=epsg:32633")} is used as UTM projection (the same projection than in Svalbard shapefiles).}

\item{proj4.deg}{desired decimal degree projection. Uses a reasonable default projection. Do not change.}

\item{round.map}{logical indicating whether the desired map is round \code{TRUE} or square \code{FALSE}. Bypasses certain routines, which caused errors.}
}
\value{
Returns a UTM coodrinate grid.
}
\description{
Outputs a list of grid lines that can be used to plot a grid on polar stereographic maps. Requires a spatial object projected using UTM coordinates
}
\seealso{
\code{\link{basemap}}
}
\author{
Mikko Vihtakari
}
\keyword{internal}
