% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_section.R
\name{interpolate_section}
\alias{interpolate_section}
\title{Interpolate x-y-z data for oceanographic section plots}
\usage{
interpolate_section(
  df = NULL,
  x,
  y,
  z,
  method = "linear",
  format = "ggplot",
  nx = 100,
  ny = 100,
  scale_distances = FALSE,
  log_y = FALSE,
  extrapolate = FALSE,
  na.rm = TRUE,
  return_df_names = FALSE
)
}
\arguments{
\item{df}{data frame. Can be \code{NULL} if \code{x, y, z} are given as numeric vectors}

\item{x, y, z}{coordinates as numeric vectors, if \code{df = NULL}. Otherwise characters specifying the column names of the coodrinates in \code{df}. The interpolation will be done as f(x, y) ~ z}

\item{method}{Character specifying the interpolation method. Options: "linear", "spline", "mba", and "idw"}

\item{format}{Character specifying the output format. Options: "ggplot" (for plotting with ggplot2), "image" (for plotting with the \code{\link[graphics]{image}} function) or "wide" for exporting as a wide data frame.}

\item{nx, ny}{Integer giving the dimensions for the output}

\item{scale_distances}{Logical indicating whether \code{x} and \code{y} should be scaled to equal distances. May be useful when interpolating variables that have different units or distances.}

\item{log_y}{logical indicating whether y axis should be interpolated using \code{log10(y + 1)} scale.}

\item{extrapolate}{logical indicating whether the interpolation function should be allowed to extrapolate beyond initial data.}

\item{na.rm}{logical indicating whether rows containing NA values should be omited from \code{df} or \code{x, y, z}. NA values are not permitted in most of the interpolation methods. See \code{\link[=oce]{sectionGrid}} for one way of interpolating NA values away from the input dataset.}

\item{return_df_names}{logical indicating whether the original column names should be returned instead of x, y, and z if \code{df != NULL}}
}
\value{
Returns a data.frame with interpolated \code{x,y,z} coordinates.
}
\description{
Wrapper function which interpolates 3D data. Designed for oceanographic section plots.
}
\details{
This is a wrapper function which uses existing interpolation solutions. Following interpolation \code{method}s have been implemented:
\itemize{
\item \code{"linear"} for linear interpolation using the \code{\link[akima]{interp}} function with \code{linear = TRUE}.
\item \code{"spline"} for spline interpolation using the \code{\link[akima]{interp}} function with \code{linear = FALSE}.
\item \code{"mba"} for multilevel B-spline interpolation using the \code{\link[MBA]{mba.surf}} function. Appears to produce the best looking results. Recommended.
\item \code{"idw"} for inverse distance weighted interpolation using the \code{\link[gstat]{idw}} function. This function was originally developed for spatial data where x- and y axes are given in the same units and have similar effect on the result. While it would, no doubt, be possible to account section design in the model specification, this has not been done in the \code{interpolate_section} function. Therefore, using this option is not recommended.
}
}
\author{
Mikko Vihtakari
}
