% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_spatial.R
\name{interpolate_spatial}
\alias{interpolate_spatial}
\title{Interpolate a spatial surface from observations}
\usage{
interpolate_spatial(
  df,
  value,
  Subset = NULL,
  coords = c("lon.utm", "lat.utm"),
  station.col = "Station",
  strata.col = "From",
  name.col = NULL,
  id.cols = NULL,
  bin.method = "average",
  int.method = "idw",
  unit = NULL,
  shear = NULL,
  n.tile = 100,
  accuracy = 100
)
}
\arguments{
\item{df}{data frame containing required information}

\item{value}{Character referring to the name of the value column to be used for interpolation.}

\item{Subset}{A subset argument as a character (i.e. with " ". See Examples).}

\item{coords}{A vector of column names for x (longitude) and y (latitude) coordinates, respectively. It is recommended to use UTM coordinates instead of decimal degrees. See \code{\link[sp]{coordinates}}.}

\item{station.col}{Character. Name of the column that specifies unique stations (i.e. spatial points). Required.}

\item{strata.col}{Character. Column that specify the sampling depth.}

\item{name.col}{Character. Column giving sample names. Not required.}

\item{id.cols}{Character vector. Identification columns that should be preserved together with value, From, To and UTM coordinate columns}

\item{bin.method}{Character giving the method for binning data, if there are several observations for each spatial point. Alternatives:
\itemize{
\item \strong{"none"} for no binning. Requires one value per unique coordinate.
\item \strong{"average"} for average values
\item \strong{"integrate"} for vertical trapezoidal integration using the \code{\link[oce]{integrateTrapezoid}} function from \link[oce]{oce} package.
}}

\item{int.method}{Character giving the method for interpolation. Currently only one option is implemented: "idw", inverse distance weighted interpolation using the \code{\link[gstat]{idw}} function.}

\item{unit}{The unit for \code{value} column. Not required.}

\item{shear}{Map tilting. Either NULL for non-tilted maps or a shear matrix, f.ex (matrix(c(2,1.2,0,1),2,2)) to shear the interpolation. This feature works poorly.}

\item{n.tile}{Number of horizontal and vertical tiles. Default is 100 resulting to 10000 tiles.}

\item{accuracy}{Number to which the extent of the interpolation area should be rounded. Given in meters.}
}
\value{
Returns a \code{spatInt} object which is a list
}
\description{
Interpolates a x-y surface from a data frame of spatial observations for plotting. Currently uses the \code{\link[gstat]{idw}} function.
}
\details{
The function removes missing values (NAs) from \code{value} column.

A word of warning about \code{bin.method = "integrate"}: this functionality works only if samples have been taken consistently at same depths. In other cases, it is recommended to use \code{bin.method = "average"}, although the user should be careful in comparing samples taken from different depths in general. The unit for integrated value is [amount]/m2, if the original value was [amount]/m3.
}
\examples{
data(chlorophyll) ## load an example dataset
x <- interpolate_spatial(chlorophyll, Subset = "From <= 10", value = "Chla") ## Interpolate
plot(x) ## Plot
}
\seealso{
\code{\link{plot.spatInt}} for plotting; \code{\link{interpolate_section}} for
}
\author{
Mikko Vihtakari
}
