% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{kong.cr}
\alias{kong.cr}
\title{SpatialPolygonsDataFrame containing clip boundaries for Kongsfjorden}
\format{SpatialPolygonsDataFrame.}
\usage{
kong.cr
}
\description{
SpatialPolygonsDataFrame containing clip boundaries for Kongsfjorden
}
\seealso{
basemap

Other shapefiles: \code{\link{arctic60}},
  \code{\link{arctic_bathy}}, \code{\link{arctic}},
  \code{\link{barents_bathy}},
  \code{\link{barents_currents}}, \code{\link{mosj.cr}},
  \code{\link{svalbard.gl}}, \code{\link{svalbard.ld}},
  \code{\link{svalbard_bathy}}
}
\author{
Mikko Vihtakari
}
\concept{shapefiles}
\keyword{datasets}
\keyword{internal}
\keyword{shapefiles}
