% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{kongsfjord_watermasses}
\alias{kongsfjord_watermasses}
\title{Water mass definitions for Kongsfjorden}
\format{data.frame}
\usage{
kongsfjord_watermasses
}
\description{
A dataframe containing water mass definitions for Kongsfjorden. An example of formatting of the \code{WM} argument in \code{\link{define_water_type}}
}
\keyword{datasets}
\keyword{ts_plot}
