% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_projection.R
\name{map_projection}
\alias{map_projection}
\title{Get the projection for a map type}
\usage{
map_projection(map.type)
}
\arguments{
\item{map.type}{A character argument giving the name of the map type. Options: "kongsfjorden", "svalbard", "barents" or "panarctic"}
}
\description{
The function prints the \code{\link[sp]{proj4string}} code used for a map type
}
\details{
Note that the map type has to be specified using the general map type name specified by the \code{$map.type} argument returned by \code{\link{map_type}}.
}
\seealso{
\code{\link{map_type}}
}
\author{
Mikko Vihtakari
}
\keyword{internal}
