% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_type.R
\name{map_type}
\alias{map_type}
\title{Get information for a map type}
\usage{
map_type(type)
}
\arguments{
\item{type}{A character argument giving the type of map area. See \code{type} in \code{\link{basemap}}.}
}
\description{
The function prints the information used in internal calculations by the \code{\link{basemap}} function
}
\seealso{
\code{\link{map_projection}} \code{\link{basemap}}
}
\author{
Mikko Vihtakari
}
\keyword{internal}
