% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{meiofauna}
\alias{meiofauna}
\title{Pan-Arctic meiofauna data from Bluhm et al. 2018}
\format{A dataframe containing pna-Arcitc meiofauna data sampled using ice-cores}
\source{
\href{http://doi.wiley.com/10.1002/ece3.3797}{Bluhm, B.A., Hop, H., Vihtakari, M., Gradinger, R., Iken, K., Melnikov, I.A., & Søreide, J.E. (2018) Sea ice meiofauna distribution on local to pan-Arctic scales. Ecology and Evolution, 8:2350-2364.}
}
\usage{
data(meiofauna)
}
\keyword{datasets}
