% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{npi_stations}
\alias{npi_stations}
\title{Location of NPI's permanent marine biological monitoring stations}
\format{
A dataframe containing following columns:
\itemize{
  \item Area. Geographic location of station.
  \item Station. Sampling station name.
  \item Lon. Longitude of the sampling location in decimal degrees
  \item Lat. Latitude of the sampling location in decimal degrees.
 }
}
\source{
Norwegian Polar Institute (\url{http://www.npolar.no/en/})
}
\usage{
data(npi_stations)
}
\description{
Location of NPI's permanent marine biological monitoring stations
}
\keyword{datasets}
