% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSvalbard.R
\docType{package}
\name{PlotSvalbard-package}
\alias{PlotSvalbard}
\alias{PlotSvalbard-package}
\title{Plot research data from Svalbard on maps}
\description{
Uses ggplot2 syntax and high resolution up-to-date shape files to plot research data from Svalbard on maps.
}
\details{
The general map-making function for \pkg{PlotSvalbard} is
\code{\link{basemap}}. This function creates a "canvas" on which research data
can be plotted. The \code{\link{basemap}} function is analogous to \code{\link[=ggplot2]{ggplot}} function
from the \pkg{ggplot2} package. Remember to use \code{data = <<NAMEOFDATASET>>} for additional geometries 
you plot on \code{basemaps} (\code{basemap("kongsfjorden") + geom_text(data = npi_stations, aes(x = lon.utm, y = lat.utm, label = Station))}
as an example).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MikkoVihtakari/PlotSvalbard}
  \item Report bugs at \url{https://github.com/MikkoVihtakari/PlotSvalbard/issues}
}

}
\author{
\strong{Maintainer}: Mikko Vihtakari \email{mikko.vihtakari@gmail.com}

Other contributors:
\itemize{
  \item Anders Skoglund [contributor]
}

}
