% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.basemapLimits.R
\name{print.basemapLimits}
\alias{print.basemapLimits}
\title{Print decimal degree grid (\code{basemapLimits}) objects}
\usage{
\method{print}{basemapLimits}(x, ...)
}
\arguments{
\item{x}{\code{basemapLimits} object to be printed.}

\item{...}{further arguments passed to \code{\link{print}}.}
}
\description{
\code{\link{print}} function for \code{\link[=basemap_limits]{basemapLimits}} objects
}
\seealso{
\code{\link{basemap_limits}} \code{\link{deg_grid}} \code{\link{basemap}}
}
\author{
Mikko Vihtakari
}
\keyword{internal}
