% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.spatInt.R
\name{print.spatInt}
\alias{print.spatInt}
\title{Print \code{spatInt} objects}
\usage{
\method{print}{spatInt}(x, ...)
}
\arguments{
\item{x}{\code{spatInt} object to be printed.}

\item{...}{further arguments passed to \code{\link{print}}.}
}
\description{
\code{\link{print}} function for \code{\link[=interpolate]{spatInt}} objects
}
\seealso{
\code{\link{interpolate}}
}
\author{
Mikko Vihtakari
}
\keyword{internal}
