% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap.R
\name{qmap}
\alias{qmap}
\title{Quick map}
\usage{
qmap(
  data,
  x = NULL,
  y = NULL,
  map.type = "panarctic",
  geom = "point",
  limits = "auto",
  bathymetry = FALSE,
  bathy.style = "poly_blues",
  ...
)
}
\arguments{
\item{data}{Data frame to use.}

\item{x, y, ...}{Aesthetics passed into each layer. Longitude and latitude columns are automatically recognized given that user has followed the PlotSvalbard standard naming. (ADD FUNCTION TO AUTOMATICALLY TRANSFORM DATA)}

\item{map.type}{Type of map area. Options: "panarctic", "barentssea", "svalbard", "mosj", "kongsfjorden", "kongsfjordbotn", "kronebreen", or "rijpfjorden". See the \code{type} argument in \code{\link{basemap}}.}

\item{geom}{Character vector specifying geom(s) to draw. Defaults to "point".}

\item{limits}{Map limits. See the \code{limits} argument in \code{\link{basemap}}. If "auto" the limits are automatically selected using the second option in the list.}

\item{bathymetry}{Logical indicating whether bathymetry should be added to the map.}

\item{bathy.style}{Character defining the style for bathymetry contours. See the \code{bathy.style} argument in \code{\link{basemap}}.}
}
\description{
\code{qmap} is a shortcut similar to ggplot2's \code{\link[ggplot2]{qplot}} designed to quickly plot data with a limited range of options.
}
\author{
Mikko Vihtakari
}
