% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{rijpfjord_watermasses}
\alias{rijpfjord_watermasses}
\title{Water mass definitions for Rijpfjorden}
\format{data.frame}
\usage{
rijpfjord_watermasses
}
\description{
A dataframe containing water mass definitions for Rijpfjorden.
}
\keyword{datasets}
\keyword{ts_plot}
