% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle_charts.R
\name{round_preserve_sum}
\alias{round_preserve_sum}
\title{Round values preserving total sums}
\usage{
round_preserve_sum(x, digits = 0)
}
\arguments{
\item{x}{numeric vector of values to be rounded}

\item{digits}{integer indicating the number of decimal places. See \code{\link[base]{round}}.}
}
\value{
Returns a numeric vector.
}
\description{
The function rounds values preserving total sums
}
\seealso{
Other waffle: 
\code{\link{waffle_chart}()},
\code{\link{waffleize}()}
}
\author{
The function is written as a communal effort. Main authors are \href{https://stackoverflow.com/questions/32544646/round-vector-of-numerics-to-integer-while-preserving-their-sum}{josliber} and \href{https://www.r-bloggers.com/round-values-while-preserve-their-rounded-sum-in-r/}{BioStatMatt}.
}
\concept{waffle}
\keyword{internal}
