% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{svalbard.ld}
\alias{svalbard.ld}
\title{SpatialPolygons object for Svalbard landshapes}
\format{Spatial polygons containing 1:250 000 landshapes of Svalbard. In \code{"+init=epsg:32633"} projection.}
\source{
Norwegian Polar Institute (\url{http://geodata.npolar.no/})
}
\usage{
svalbard.ld
}
\description{
SpatialPolygons object for Svalbard landshapes
}
\keyword{datasets}
