% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{svalbard_bathy}
\alias{svalbard_bathy}
\title{SpatialPolygonsDataFrame containing detailed bathymetry for Svalbard region}
\format{SpatialPolygonsDataFrame}
\source{
\href{https://kartkatalog.geonorge.no/metadata/kartverket/sjokart-dybdedata/2751aacf-5472-4850-a208-3532a51c529a}{Norwegian Mapping Authority}
}
\usage{
svalbard_bathy
}
\description{
SpatialPolygonsDataFrame containing detailed bathymetry for Svalbard region
}
\seealso{
basemap

Other shapefiles: \code{\link{arctic50}},
  \code{\link{arctic60}}, \code{\link{arctic_bathy}},
  \code{\link{barents_bathy}},
  \code{\link{barents_currents}}, \code{\link{kong.cr}},
  \code{\link{mosj.cr}}, \code{\link{svalbard.gl}},
  \code{\link{svalbard.ld}}
}
\concept{shapefiles}
\keyword{datasets}
\keyword{internal}
\keyword{shapefiles}
