% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle_charts.R
\name{waffle_chart}
\alias{waffle_chart}
\title{Plot waffle charts}
\usage{
waffle_chart(data, fill, value = "value", facet = NULL,
  composition = TRUE, max_value = NULL, digits = 3,
  fill_colors = NULL, fill_title = NULL, ncol = NULL,
  base_size = 12, line_size = LS(1), legend.position = "bottom")
}
\arguments{
\item{data}{data frame to be plotted}

\item{fill}{character specifying the column name which should be used as fill for the waffle plot.}

\item{value}{character specifying the column name which contains values of the \code{fill} variable. Will be used to fill the waffle cells.}

\item{facet}{character specifying the column name which should be used to \code{\link[ggplot2]{facet_wrap}} waffle charts.}

\item{composition}{logical indicating whether a compositional waffle (i.e. fill adds up to 100\%) should be created. If \code{FALSE}, waffle cells will be scaled to \code{max_value} and missing cells filled with an "empty" category.}

\item{max_value}{numerical giving the value to which waffle cells should be scaled to, if \code{composition = FALSE}.}

\item{digits}{integer indicating the number of decimal places to be used in rounding of the waffle cells. The value 3 indicates percentages, while 4 permilles.}

\item{fill_colors}{named character vector giving the colors for \code{fill} levels. See \code{\link[ggplot2]{scale_fill_manual}}.}

\item{fill_title}{character giving the title for the color legend.}

\item{ncol}{number of columns to be used in facetting. See \code{\link[ggplot2]{facet_wrap}}.}

\item{base_size}{numeric giving the base size for the plot. See \code{\link[ggplot2]{theme_void}}.}

\item{line_size}{numeric value giving the size of lines between waffles cells.}

\item{legend.position}{character specifying the position of the legend. See \code{\link[ggplot2]{ggtheme}}.}
}
\value{
Returns a \link[ggplot2]{ggplot2} waffle plot
}
\description{
The function uses \link[ggplot2]{ggplot2} to create waffle charts from data.
}
\details{
The waffle charts are read from left to right (like text) and from bottom upwards (like water glass). The cells indicate 1\% of the maximum value (100% if \code{composition = TRUE} else \code{max_value}). The cells are divided vertically to fractions specifies by the \code{digits} argument.
}
\seealso{
Other waffle: \code{\link{round_preserve_sum}},
  \code{\link{waffleize}}
}
\author{
Mikko Vihtakari with code ideas from \href{https://github.com/hrbrmstr/waffle}{hrbrmstr} and \href{https://github.com/liamgilbey/ggwaffle}{Liam Gilbey}
}
\concept{waffle}
