% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{arctic_bathy}
\alias{arctic_bathy}
\title{SpatialPolygonsDataFrame containing pan-Arctic bathymetry}
\format{SpatialPolygonsDataFrame containing pan-Arctic bathymetry. In polar stereographic 71 degrees projection.}
\source{
\href{https://www.ngdc.noaa.gov/mgg/bathymetry/arctic/ibcaoversion3.html}{Jakobsson, M., et al. The International Bathymetric Chart of the Arctic Ocean (IBCAO) Version 3.0. Geophys. Res. Lett. 2012, 39:L12609}
}
\usage{
arctic_bathy
}
\description{
SpatialPolygonsDataFrame containing pan-Arctic bathymetry
}
\seealso{
basemap

Other shapefiles: \code{\link{arctic50}},
  \code{\link{arctic60}}, \code{\link{arctic_currents}},
  \code{\link{atlantic_currents}},
  \code{\link{barents_bathy}}, \code{\link{kong.cr}},
  \code{\link{mosj.cr}}, \code{\link{svalbard.gl}},
  \code{\link{svalbard.ld}}, \code{\link{svalbard_bathy}}
}
\concept{shapefiles}
\keyword{datasets}
\keyword{internal}
\keyword{shapefiles}
