% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{basemap}
\alias{basemap}
\title{Create a ggplot2 basemap for plotting variables}
\source{
Svalbard shape files originate from the Norwegian Polar Institute (\url{http://geodata.npolar.no/}). Barents Sea map is downloaded from  \url{http://www.naturalearthdata.com} and uses the \code{ne_10m_land} (v 4.0.0) dataset.
}
\usage{
basemap(type = "kongsfjorden", land.col = "#eeeac4", gla.col = "grey95",
  grid.col = "grey70", limits = NULL, round.lat = FALSE, n.lat.grid = 3,
  lat.interval = 10, lon.interval = 45, round.lon = FALSE,
  n.lon.grid = 3, keep.glaciers = TRUE, size.land = 0.1,
  size.glacier = 0.1, size.grid = 0.1, border.col.land = "black",
  border.col.glacier = "black")
}
\arguments{
\item{type}{Type of map area. Options: "svalbard", "mosj", "kongsfjorden", "kongsfjordbotn", "kronebreen", "barentssea", "arctic50" or "arctic60". See details.}

\item{land.col}{Color of land}

\item{gla.col}{Color of glaciers}

\item{grid.col}{Color of grid lines. Use \code{NA} to remove the grid lines.}

\item{limits}{Map limits. A numeric vector of length 4 where first element defines the minimum longitude, second element the maximum longitude, third element the minimum latitude and fourth element the maximum latitude of the bounding box. The coordinates have to be given as decimal degrees.}

\item{round.lat}{specifying the level of rounding to be used to plot latitude grid lines. Overrides \code{n.lat.grid}}

\item{n.lat.grid}{number of latitude grid lines. Alternatively use \code{round.lat}}

\item{lat.interval}{the interval of latitude grids for polar stereographic maps (\code{type = "arctic50"} or \code{"arctic60"})}

\item{lon.interval}{the interval of longitude grids for polar stereographic maps (\code{type = "arctic50"} or \code{"arctic60"})}

\item{round.lon}{specifying the level of rounding to be used to plot longitude grid lines. Overrides \code{n.lon.grid}}

\item{n.lon.grid}{number of longitude grid lines. Alternatively use \code{round.lon}}

\item{keep.glaciers}{a logical indicating whether glaciers should be kept for the Svalbard maps. Setting this to \code{FALSE} speeds up map plotting by a few seconds.}

\item{size.land}{width of the border line for land shapes. See details for explanation about line widths.}

\item{size.glacier}{width of the border line for glacier shapes.}

\item{size.grid}{width of the grid lines.}

\item{border.col.land}{color of the border line for land shapes.}

\item{border.col.glacier}{color of the border line for glacier shapes.}
}
\value{
Returns a \link[ggplot2]{ggplot2} map, which can be assigned to an object and modified as any ggplot object.
}
\description{
Creates a ggplot2 basemap for further plotting of variables.
}
\details{
The function uses \link[ggplot2]{ggplot2} and up-to-date (2017) detailed shapefiles to plot maps of Svalbard and other polar regions. The map type is defined using the \code{type} argument and map limits can be controlled with the \code{limits} argument. Currently implemented map \code{type}s:
\itemize{
\item "svalbard". Detailed 1:250 000 map of Svalbard land and glaciers. This option is slow (approx. 25 seconds) due to the large file size.
\item "mosj" shows Kongsfjoden and Fram Strait as sampled during Norwegian Polar Institute's MOSJ campaigns. Some glaciers can be older than 2015.
\item "kongsfjorden" shows Kongsfjorden and parts of Prins Karls Forland. Glaciers are from 2015 to 2017. This map uses a subset of the Svalbard shape files and is faster to plot due to smaller file size.
\item "kongsfjordbotn" shows Kongsvegen, Kronebreen, Kongsbreen and Conwaybreen. Glaciers are from July 2017. This map uses a subset of the Svalbard shape files and is faster to plot due to a smaller file size.
\item "kronebreen" shows mostly Kronebreen and Kongsvegen. Glacier fronts are from July 2017. This map uses a subset of the Svalbard shape files and is faster to plot due to a smaller file size.
\item "barentssea". A 1:10 000 000 map of the Barents Sea. See \strong{Source}.
\item "arctic50". A polar stereographic map of the Arctic with a limit at 50 degrees North.
\item "arctic60". A polar stereographic map of the Arctic with a limit at 60 degrees North.
}

Svalbard and Barents Sea maps use the \code{"+init=epsg:32633"} UTM projection. More maps can be added based on need.

\strong{Line width} (size) aesthatics in \link[ggplot2]{ggplot2} generetes approximately 2.13 wider lines measured in pt than the given values. If you want a specific line width in pt, multiply it by 1/2.13.
}
\examples{
basemap() ## Plots Kongsfjorden

## Maps work as normal ggplot2 objects:

data(chlorophyll)
p <- basemap("mosj")
p + geom_point(data = chlorophyll, aes(x = lon.utm, y = lat.utm,
size = Chla, color = Chla), shape = 1)

## limitting maps is possible, but you might need to define grid lines by hand:
basemap("kongsfjordbotn", limits = c(12.2,12.65,78.95,79.00))
basemap("kongsfjordbotn", limits = c(12.2,12.65,78.95,79.00),
round.lat = 0.01, round.lon = 0.1) # better

## Svalbard map. Warning: this is SLOW
basemap("svalbard")

## grid.col = NA removes grid lines
basemap("svalbard", grid.col = NA, limits = c(10, 28, 79.5, 83))

## Barents Sea
basemap("barentssea") # note wrong placing of axis labels

## Barents Sea map also prints mainland Norway,
## but the projection is not optimal.
basemap("barentssea", limits = c(12, 24, 68, 71))

}
\seealso{
\code{\link[ggplot2]{ggplot2}} \code{\link{theme_map}}

\code{coastlineWorldMedium} from the \code{oce} package for plotting maps in maps in base graphics using the \code{ne_10m_land} dataset.
}
\author{
Mikko Vihtakari
}
