% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_bathymetry.R
\name{clip_bathymetry}
\alias{clip_bathymetry}
\title{Clip a bathymetry shapefile to fit a basemap}
\usage{
clip_bathymetry(X)
}
\arguments{
\item{X}{a basemapData object}
}
\description{
An internal utility function to clip bathymetry to fit the boundaries of a basemap
}
\author{
Mikko Vihtakari
}
\keyword{internal}
