% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{ctd_kongsfjord}
\alias{ctd_kongsfjord}
\title{Example CTD data for Kongsfjorden}
\format{
data.frame
}
\usage{
ctd_kongsfjord
}
\description{
A dataframe containing example CTD data for Kongsfjorden.
}
\seealso{
Other ts_plot: 
\code{\link{define_water_type}()},
\code{\link{kongsfjord_watermasses}},
\code{\link{rijpfjord_watermasses}},
\code{\link{ts_plot}()}
}
\concept{ts_plot}
\keyword{datasets}
