% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{ctd_rijpfjord}
\alias{ctd_rijpfjord}
\title{Example CTD data for Rijpfjorden}
\format{
data.frame
}
\usage{
ctd_rijpfjord
}
\description{
A dataframe containing an example CTD transect data from Rijpfjorden.
}
\concept{ts_plot section_plot}
\keyword{datasets}
