% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_grid.R
\name{deg_grid}
\alias{deg_grid}
\title{Decimal degree grid from UTM projected spatial objects}
\usage{
deg_grid(dat, round.lat = FALSE, n.lat.grid = 3, round.lon = FALSE,
  n.lon.grid = 3, n.points = 10, proj4.utm = TRUE,
  proj4.deg = "+proj=longlat +datum=WGS84")
}
\arguments{
\item{dat}{Spatial object projected using UTM coordinates, such as \code{\link{readShapeSpatial}} OR a numeric vector of length 4 where first element defines the minimum longitude, second element the maximum longitude, third element the minimum latitude and fourth element the maximum latitude of the bounding box. In case of a numeric vector, the coordinates have to be given as decimal degrees.}

\item{round.lat}{specifying the level of rounding to be used to plot latitude grid lines. Overrides \code{n.lat.grid}}

\item{n.lat.grid}{number of latitude grid lines. Alternatively use \code{round.lat}}

\item{round.lon}{specifying the level of rounding to be used to plot longitude grid lines. Overrides \code{n.lon.grid}}

\item{n.lon.grid}{number of longitude grid lines. Alternatively use \code{round.lon}}

\item{n.points}{number of points used in creating the grid lines. The more points, the smoother the lines.}

\item{proj4.utm}{projection in \code{dat}. If TRUE (default), fetches the projection from \code{dat}. If \code{dat} is a numeric vector and \code{proj4.utm = TRUE}, \code{CRS("+init=epsg:32633")} is used as UTM projection (the same projection than Kongsfjorden shapefiles).}

\item{proj4.deg}{desired decimal degree projection. Uses a reasonable default projection. Do not change.}
}
\value{
Returns a UTM coodrinate grid.
}
\description{
Outputs a list of grid lines that can be used to plot a grid on maps. Requires a spatial object projected using UTM coordinates
}
\author{
Mikko Vihtakari
}
