% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spatInt.R
\name{plot.spatInt}
\alias{plot.spatInt}
\title{Plot interpolated map of Kongsfjorden}
\usage{
plot.spatInt(x, basemap = TRUE, type = "kongsfjorden",
  col.scale.limits = NULL, legend.label = NA, limits = NULL,
  round.lat = FALSE, n.lat.grid = 3, round.lon = FALSE, n.lon.grid = 3,
  land.col = "#eeeac4", gla.col = "grey95", grid.col = "grey70", ...)
}
\arguments{
\item{x}{\code{spatInt} object from \code{\link{interpolate}} function.}

\item{basemap}{Logical. Should basemap from \code{\link{basemap}} be used? Not implemented}

\item{type}{Type of basemap. See \code{\link{basemap}}. Used only if \code{basemap = TRUE}.}

\item{legend.label}{Label for color legend. If NA (default), the labels are extracted from the \code{spatInt} object.}

\item{limits}{Map limits. See \code{\link{basemap}}}

\item{...}{Additional arguments passed somewhere.}

\item{col.scale.lims}{A numeric vector of lenght 2 defining the limits for color scale of interpolated results. If \code{NULL} (default), the limits will be generated automatically.}
}
\description{
Plot method for \code{\link[=interpolate]{spatInt}} objects.
}
\examples{
data(chlorophyll) ## load an example dataset
x <- interpolate(chlorophyll, Subset = "From <= 10", value = "Chla") ## Interpolate
plot(x, type = "kongsfjorden") ## Plot

## PlotSvalbard functions can be expanded by using ggplot2 syntax
plot(x, type = "kongsfjorden") + geom_text(data = chlorophyll,
aes(x = lon.utm, y = lat.utm, label = Station))

## Changing limits is easiest using the inbuild argument
plot(x, type = "kongsfjorden", limits = c(11.4,12.7,78.85,79.05))
}
\seealso{
\code{\link{interpolate}}
}
\author{
Mikko Vihtakari
}
