% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spatInt.R
\name{plot.spatInt}
\alias{plot.spatInt}
\title{Plot an interpolated map}
\usage{
\method{plot}{spatInt}(x, basemap = TRUE, type = "kongsfjorden",
  col.scale.limits = NULL, legend.label = NA, limits = NULL,
  round.lat = FALSE, n.lat.grid = 3, round.lon = FALSE, n.lon.grid = 3,
  keep.glaciers = TRUE, land.col = "#eeeac4", size.land = 0.1,
  border.col.land = "black", gla.col = "grey95", size.glacier = 0.1,
  border.col.glacier = "black", grid.col = "grey70", size.grid = 0.1, ...)
}
\arguments{
\item{x}{\code{spatInt} object from \code{\link{interpolate}} function.}

\item{basemap}{Logical. Should basemap from \code{\link{basemap}} be used? Not implemented}

\item{type}{Type of basemap. See \code{\link{basemap}}. Used only if \code{basemap = TRUE}.}

\item{col.scale.limits}{A numeric vector of lenght 2 defining the limits for color scale of interpolated results. If \code{NULL} (default), the limits will be generated automatically.}

\item{legend.label}{Label for color legend. If NA (default), the labels are extracted from the \code{spatInt} object.}

\item{limits}{Map limits. See \code{\link{basemap}}}

\item{round.lat}{specifying the level of rounding to be used to plot latitude grid lines. Overrides \code{n.lat.grid}}

\item{n.lat.grid}{number of latitude grid lines. Alternatively use \code{round.lat}}

\item{round.lon}{specifying the level of rounding to be used to plot longitude grid lines. Overrides \code{n.lon.grid}}

\item{n.lon.grid}{number of longitude grid lines. Alternatively use \code{round.lon}}

\item{keep.glaciers}{a logical indicating whether glaciers should be kept for the Svalbard maps. Setting this to \code{FALSE} speeds up map plotting little bit.}

\item{land.col}{Color of land}

\item{gla.col}{Color of glaciers}

\item{grid.col}{Color of grid lines. Use \code{NA} to remove the grid lines.}

\item{...}{Additional arguments passed somewhere.}
}
\description{
Plot method for \code{\link[=interpolate]{spatInt}} objects.
}
\examples{
data(chlorophyll) ## load an example dataset
x <- interpolate(chlorophyll, Subset = "From <= 10", value = "Chla") ## Interpolate
plot(x, type = "kongsfjorden") ## Plot

## PlotSvalbard functions can be expanded by using ggplot2 syntax
plot(x, type = "kongsfjorden") + geom_text(data = chlorophyll,
aes(x = lon.utm, y = lat.utm, label = Station))

## Changing limits is easiest using the inbuild argument
plot(x, type = "kongsfjorden", limits = c(11.4,12.7,78.85,79.05))
}
\seealso{
\code{\link{interpolate}}
}
\author{
Mikko Vihtakari
}
