% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spatInt.R
\name{plot.spatInt}
\alias{plot.spatInt}
\title{Plot an interpolated map}
\usage{
\method{plot}{spatInt}(
  x,
  basemap = TRUE,
  type = "svalbard",
  limits = "auto",
  limits.lon = 0.001,
  limits.lat = 0.001,
  col.scale.limits = NULL,
  legend.label = NA,
  round.lat = FALSE,
  n.lat.grid = 3,
  round.lon = FALSE,
  n.lon.grid = 3,
  keep.glaciers = TRUE,
  land.col = "grey60",
  land.size = 0.1,
  land.border.col = "black",
  gla.col = "grey95",
  gla.size = 0.1,
  gla.border.col = "black",
  grid.col = "grey70",
  grid.size = 0.1,
  base_size = 11,
  legend.position = "right",
  ...
)
}
\arguments{
\item{x}{\code{spatInt} object from \code{\link{interpolate_spatial}} function.}

\item{basemap}{Logical. Should basemap from \code{\link{basemap}} be used? Not implemented}

\item{type}{Type of basemap. See \code{\link{basemap}}. Used only if \code{basemap = TRUE}.}

\item{limits}{Map limits. See \code{\link{basemap}}. The option "auto" (default) limits the map using coordinate range in \code{x}. Replace by \code{NULL} to remove the automatic zooming. Alternatively use a numeric vector as described in \code{\link{basemap}} documentation.}

\item{limits.lon, limits.lat, round.lon, round.lat, n.lon.grid, n.lat.grid, land.size, land.col, gla.col, grid.col, land.border.col, gla.size, gla.border.col, grid.size, base_size, keep.glaciers, legend.position}{See \code{\link{basemap}} documentation.}

\item{col.scale.limits}{A numeric vector of lenght 2 defining the limits for color scale of interpolated results. If \code{NULL} (default), the limits will be generated automatically.}

\item{legend.label}{Label for color legend. If NA (default), the labels are extracted from the \code{spatInt} object.}

\item{...}{Additional arguments. Required by R build checks. Ignore.}
}
\description{
Plot method for \code{\link[=interpolate_spatial]{spatInt}} objects.
}
\examples{
data(chlorophyll) ## load an example dataset
x <- interpolate_spatial(chlorophyll, Subset = "From <= 10", value = "Chla") ## Interpolate
plot(x) ## Plot

## PlotSvalbard functions can be expanded by using ggplot2 syntax
plot(x) + geom_text(data = chlorophyll, aes(x = lon.utm, y = lat.utm, label = Station))

## Auto limits leave an empty space around the gridded surface.
## Changing limits is easiest using the inbuild argument
plot(x, limits = c(11.4,12.7,78.85,79.05))
}
\seealso{
\code{\link{interpolate_spatial}} \code{\link{basemap}}
}
\author{
Mikko Vihtakari
}
