% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_plot.R
\name{section_plot}
\alias{section_plot}
\title{Make an oceanographic section plot from an x-y-z data frame}
\usage{
section_plot(df, x, y, z, bottom = NULL, interpolate = FALSE,
  interp_method = "mba", log_y = FALSE, xlab = "Distance",
  ylab = "Depth", zlab = "Variable", ybreaks = waiver(),
  xbreaks = waiver(), zbreaks = waiver(), xlim = NULL, ylim = NULL,
  zlim = NULL, zrange = c(1, 6), add_bottom = NULL,
  legend.position = "right", base_size = 10)
}
\arguments{
\item{df}{Data frame containing the data}

\item{x}{Character specifying the column name which should be used as x (=distance) axis}

\item{y}{Character specifying the column name which should be used as y (=depth) axis. The y axis will be inverted in the resulting plot.}

\item{z}{Character specifying the column name which should be for bubbles or interpolation fill depending on the \code{interpolate} argument.}

\item{bottom}{Optional character specifying the column name which contains bottom depths OR a data frame with x-axis values in the first column and bottom depths in the second column. If provided, unique combinations of \code{x} and \code{bottom} will be used to plot bottom for the section plot}

\item{interpolate}{Logical indicating whether an interpolated section plot (\code{TRUE}) or a bubble plot (\code{FALSE}) should be returned.}

\item{interp_method}{Character specifying the interpolation method, if \code{interpolate = TRUE}. See \code{\link{interpolate_section}}.}

\item{log_y}{logical indicating whether the y-axis should be \code{log10(y + 10)} transformed before plotting. Helps showing differences close to the surface if some stations are much deeper than others.}

\item{xlab, ylab, zlab}{Character specifying the labels for x-axis, y-axis and legend, respectively.}

\item{xbreaks, ybreaks}{Numeric vector specifying the breaks for x and y-axes. See \link[ggplot]{scale_continuous}.}

\item{zbreaks}{Numeric vector specifying the breaks for z legend if \code{interpolate = FALSE} or contour lines if  \code{interpolate = TRUE}}

\item{xlim, ylim}{Numeric vector of length two providing limits of the scale. Use NA to refer to the existing minimum or maximum. Use \code{NULL} for default limits.}

\item{zlim}{Numeric vector of length two providing limits for fill or bubble size. Any values outside these limits will get the extreme values defined by \code{zrange}}

\item{zrange}{Range for z graphical parameter. At the moment only considered for \code{interpolate = FALSE}.}

\item{add_bottom}{A numeric vector of length two providing the depths that should be added to \code{bottom} at the extremes (\code{\link[base]{range}}) of \code{xbreaks}.}

\item{legend.position}{Position for ggplot2 legend. See the argument with the same name in\link[ggplot2]{theme}.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{theme_bw}.}
}
\description{
Make an oceanographic section plot from an x-y-z data frame
}
\details{
Creates an oceanographic section plot from a data frame using ggplot2. Returns either an interpolated section or a bubble plot.
}
\author{
Mikko Vihtakari
}
