% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select an element of each vector from a list}
\usage{
select(x, y)
}
\arguments{
\item{x}{list}

\item{y}{number of element. Must be numeric}
}
\description{
Selects y'th element of each vector from a list
}
