% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{svalbard.ld}
\alias{svalbard.ld}
\title{SpatialPolygonsDataFrame containing land shapes for Svalbard}
\format{
SpatialPolygonsDataFrame containing 1:250 000 landshapes of Svalbard. In \code{"+init=epsg:32633"} projection.
}
\source{
Norwegian Polar Institute (\url{http://geodata.npolar.no/})
}
\usage{
svalbard.ld
}
\description{
SpatialPolygonsDataFrame containing land shapes for Svalbard
}
\seealso{
basemap

Other shapefiles: 
\code{\link{arctic60}},
\code{\link{arctic_bathy}},
\code{\link{arctic}},
\code{\link{barents_bathy}},
\code{\link{barents_currents}},
\code{\link{kong.cr}},
\code{\link{mosj.cr}},
\code{\link{svalbard.gl}},
\code{\link{svalbard_bathy}}
}
\concept{shapefiles}
\keyword{datasets}
\keyword{internal}
\keyword{shapefiles}
